Added behaviors:
altspaceutil.behaviors.HoverMaterialOpacity
Changes the opacity of an object's material when the cursor hovers over it, and restores the original opacity when the cursor is no longer hovering over the object.

altspaceutil.behaviors.HoverMaterialColor
Changes the color of an object's material when the cursor hovers over it, and restores the original color when the cursor is no longer hovering over the object.

altspaceutil.behaviors.NativeTextMaterial
Updates the color and opacity of a n-text native component using a material source.